+++ CHANGELOG.md
@@ -4,6 +4,16 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## `@m2c2kit/core` [0.3.2] - 2022-10-26
+
+### Added
+
+- Get version string from `package.json` file.
+
+### Fixed
+
+- Tests
+
 ## `@m2c2kit/core` [0.3.1] - 2022-10-25
 
 ### Added